refactor: MeasurementVC에서 View 정의하도록 변경 #2
- MeasurementView 삭제
